pyproject.toml
[project]
name = "derma-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "faiss-cpu==1.7.4",
    "fastapi==0.104.1",
    "langchain==0.3.26",
    "langchain-community==0.3.26",
    "langchain-openai==0.3.24",
    "langchain-pinecone==0.2.8",
    "numpy>=1.26.4",
    "pydantic-settings>=2.4.0",
    "pypdf==5.6.1",
    "python-dotenv==1.1.0",
    "sentence-transformers==4.1.0",
    "uvicorn[standard]==0.24.0",
]
